Test i fixos fets

+++ pooparking/src/main/java/dam/com/Places/Coordenada.java
@@ -5,7 +5,7 @@ public class Coordenada {
     private double y;
 
     public Coordenada(double x, double y) {
-        
+
         this.x = x;
         this.y = y;
 
@@ -15,7 +15,7 @@ public class Coordenada {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -23,10 +23,8 @@ public class Coordenada {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
-
-
 
 }
